changing ad check to hook
